perf:优化积分、消息框的样式
